Refactoring.

+++ debug.c
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 #include <raylib.h>
 #include <box2d/box2d.h>
 
@@ -10,7 +12,7 @@ void draw_segment(b2Vec2 p1, b2Vec2 p2, b2HexColor color, void* context) {
 }
 
 void draw_polygon(const b2Vec2* vertices, int vertexCount, b2HexColor color, void* context) {
-	for (int i = 0; i < vertexCount - 1; i++) {
+	for (size_t i = 0; i < vertexCount - 1; i++) {
 		draw_segment(vertices[i], vertices[i + 1], color, context);
 	}
 
@@ -19,7 +21,7 @@ void draw_polygon(const b2Vec2* vertices, int vertexCount, b2HexColor color, voi
 
 void draw_solid_polygon(b2Transform transform, const b2Vec2* vertices, int vertexCount, float radius,
 	b2HexColor color, void* context) {
-    for (int i = 0; i < vertexCount; ++i) {
+    for (size_t i = 0; i < vertexCount; ++i) {
         int next_index = (i + 1 == vertexCount) ? 0 : i + 1;
         b2Vec2 p0 = b2TransformPoint(transform, vertices[i]);
         b2Vec2 p1 = b2TransformPoint(transform, vertices[next_index]);
